Working on splitting up code however this wont work as the after-effects module doesnt allow for building and passing js files :( Need to figure out another way to automate running scripts in AE
