+++ .gitignore
@@ -3,3 +3,5 @@ node_modules
 *.log
 .DS_Store
 bundle.js
+/index.js
+/test/aeToJSON.js
