+++ .npmignore
@@ -8,3 +8,4 @@ test.js
 demo/
 .npmignore
 LICENSE.md
+/src
