+++ package.json
@@ -16,16 +16,19 @@
     "after-effects": "^0.3.2",
     "babel-cli": "^6.6.5",
     "babel-preset-es2015": "^6.6.0",
-    "opn": "^4.0.1",
     "tape": "^4.5.1"
   },
   "scripts": {
-    "test": "node test/",
-    "prePublish": "npm run build-jsx",
-    "build-jsx": "rm bundle.jsx; npm run copy-shim && npm run copy-get && npm run js-jsx",
-    "copy-shim": "cat node_modules/after-effects/lib/includes/es5-shim.jsx >> bundle.jsx",
-    "copy-get": "cat node_modules/after-effects/lib/includes/get.jsx >> bundle.jsx",
-    "js-jsx": "browserify index.jsx >> bundle.jsx; babel bundle.jsx >> bundle.jsx"
+    "test": "npm run transpile; npm run browserify-test; npm run clean-transpile; babel-node test/",
+    "browserify-test": "browserify index.js -s aeToJSON -o test/aeToJSON.js",
+    "prepublish": "npm run compile",
+    "postpublish": "npm run clean-transpile",
+    "compile": "npm run transpile && npm run browserify-jsx",
+    "clean-transpile": "rm *.js *.jsx; rm -rf shims/",
+    "transpile": "babel -x .js -d ./ src/; babel src/index.jsx -o index.jsx",
+    "browserify-jsx": "browserify index.jsx -o aeToJSON.jsx; rm index.jsx",
+    "copy-shim": "cat node_modules/after-effects/lib/includes/es5-shim.jsx > bundle.jsx",
+    "copy-get": "cat node_modules/after-effects/lib/includes/get.jsx > bundle.jsx"
   },
   "keywords": [
     "after,effects,json,jsx"
