+++ src/getLayer.js
@@ -1,0 +1,8 @@
+const getProperties = require('./getProperties');
+
+module.exports = function getLayer(layer) {
+  return {
+    name: layer.name,
+    properties: getProperties(layer)
+  };
+};
