+++ src/index.js
@@ -1,0 +1,7 @@
+module.exports = function() {
+  const getCompositions = require('./getCompositions');
+  
+  return {
+    compositions: getCompositions()
+  };
+};
